Do not use ATC to stop Logging_Task in active logger
Using abort and ATC, which is basically an implicit abort, to stop tasks
is problematic because *any* data referenced by a task after the abort
is in potentially abnormal state.

Use a special log request to signal termination instead. This allows us
to get rid of the trigger type as well.