Delegate install and gpr generation to gprinstall
The Makefile depends less on the build directory structure while the
installation is fully configurable via DESTDIR, PREFIX and
GPRINSTALLFLAGS.

The installed sources now include C sources. This seems good for
debugging, and the intent has never been to allow a rebuild.

The generated project contains lots of unneeded variables, but this
does not hurt and may arguably be useful.

Remove LIBRARY_VERSION Make variable.